+++ AltairHL_emulator/PortDrivers/utility_io.h
@@ -8,5 +8,5 @@
 
 extern const char ALTAIR_EMULATOR_VERSION[];
 
-size_t utility_output(int port, int data, char *buffer, size_t buffer_length);
+size_t utility_output(int port, uint8_t data, char *buffer, size_t buffer_length);
 uint8_t utility_input(uint8_t port);
